Suppress warnings

+++ src/data_store_kmr.cpp
@@ -830,8 +830,8 @@ namespace kmrnext {
     // Search data should be sent
     vector< vector<DataPack> > dpgroups(ndata);
     {
-      size_t range_start = (indices_cnt > 0)? indices[0] : -1;
-      size_t range_end   = (indices_cnt > 0)? indices[indices_cnt - 1] : -1;
+      long range_start = (indices_cnt > 0)? indices[0] : -1;
+      long range_end   = (indices_cnt > 0)? indices[indices_cnt - 1] : -1;
 #ifdef _OPENMP
       #pragma omp parallel for
 #endif
@@ -845,7 +845,7 @@ namespace kmrnext {
 	}
 	data_[i].unshared();
 	Key tmpkey = index_to_key(i);
-	size_t viewed_idx = key_to_viewed_index(tmpkey, aview);
+	long viewed_idx = (long)key_to_viewed_index(tmpkey, aview);
 	if (!(range_start <= viewed_idx && viewed_idx <= range_end)) {
 	  vector<DataPack>& dps = dpgroups.at(viewed_idx);
 #ifdef _OPENMP
